тестовый пример для .net и com dll
