--- DesktopEditor/doctrenderer/test_builder/docbuilder/docbuilder.sln
@@ -1,61 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 2013
-VisualStudioVersion = 12.0.30723.0
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "doctrendererwrapper", "doctrendererwrapper\doctrendererwrapper.vcxproj", "{10124551-28B8-4CA0-8FBA-420CF9602CF3}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "test", "test\test.csproj", "{85EE7B43-D2BF-4E8B-A103-7D1845A21587}"
-	ProjectSection(ProjectDependencies) = postProject
-		{10124551-28B8-4CA0-8FBA-420CF9602CF3} = {10124551-28B8-4CA0-8FBA-420CF9602CF3}
-	EndProjectSection
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Debug|ARM = Debug|ARM
-		Debug|Mixed Platforms = Debug|Mixed Platforms
-		Debug|Win32 = Debug|Win32
-		Debug|x64 = Debug|x64
-		Release|Any CPU = Release|Any CPU
-		Release|ARM = Release|ARM
-		Release|Mixed Platforms = Release|Mixed Platforms
-		Release|Win32 = Release|Win32
-		Release|x64 = Release|x64
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{10124551-28B8-4CA0-8FBA-420CF9602CF3}.Debug|Any CPU.ActiveCfg = Debug|Win32
-		{10124551-28B8-4CA0-8FBA-420CF9602CF3}.Debug|ARM.ActiveCfg = Debug|Win32
-		{10124551-28B8-4CA0-8FBA-420CF9602CF3}.Debug|Mixed Platforms.ActiveCfg = Debug|Win32
-		{10124551-28B8-4CA0-8FBA-420CF9602CF3}.Debug|Mixed Platforms.Build.0 = Debug|Win32
-		{10124551-28B8-4CA0-8FBA-420CF9602CF3}.Debug|Win32.ActiveCfg = Debug|Win32
-		{10124551-28B8-4CA0-8FBA-420CF9602CF3}.Debug|Win32.Build.0 = Debug|Win32
-		{10124551-28B8-4CA0-8FBA-420CF9602CF3}.Debug|x64.ActiveCfg = Debug|x64
-		{10124551-28B8-4CA0-8FBA-420CF9602CF3}.Debug|x64.Build.0 = Debug|x64
-		{10124551-28B8-4CA0-8FBA-420CF9602CF3}.Release|Any CPU.ActiveCfg = Release|Win32
-		{10124551-28B8-4CA0-8FBA-420CF9602CF3}.Release|ARM.ActiveCfg = Release|Win32
-		{10124551-28B8-4CA0-8FBA-420CF9602CF3}.Release|Mixed Platforms.ActiveCfg = Release|Win32
-		{10124551-28B8-4CA0-8FBA-420CF9602CF3}.Release|Mixed Platforms.Build.0 = Release|Win32
-		{10124551-28B8-4CA0-8FBA-420CF9602CF3}.Release|Win32.ActiveCfg = Release|Win32
-		{10124551-28B8-4CA0-8FBA-420CF9602CF3}.Release|Win32.Build.0 = Release|Win32
-		{10124551-28B8-4CA0-8FBA-420CF9602CF3}.Release|x64.ActiveCfg = Release|Win32
-		{85EE7B43-D2BF-4E8B-A103-7D1845A21587}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{85EE7B43-D2BF-4E8B-A103-7D1845A21587}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{85EE7B43-D2BF-4E8B-A103-7D1845A21587}.Debug|ARM.ActiveCfg = Debug|Any CPU
-		{85EE7B43-D2BF-4E8B-A103-7D1845A21587}.Debug|Mixed Platforms.ActiveCfg = Debug|Any CPU
-		{85EE7B43-D2BF-4E8B-A103-7D1845A21587}.Debug|Mixed Platforms.Build.0 = Debug|Any CPU
-		{85EE7B43-D2BF-4E8B-A103-7D1845A21587}.Debug|Win32.ActiveCfg = Debug|Any CPU
-		{85EE7B43-D2BF-4E8B-A103-7D1845A21587}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{85EE7B43-D2BF-4E8B-A103-7D1845A21587}.Debug|x64.Build.0 = Debug|Any CPU
-		{85EE7B43-D2BF-4E8B-A103-7D1845A21587}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{85EE7B43-D2BF-4E8B-A103-7D1845A21587}.Release|Any CPU.Build.0 = Release|Any CPU
-		{85EE7B43-D2BF-4E8B-A103-7D1845A21587}.Release|ARM.ActiveCfg = Release|Any CPU
-		{85EE7B43-D2BF-4E8B-A103-7D1845A21587}.Release|Mixed Platforms.ActiveCfg = Release|Any CPU
-		{85EE7B43-D2BF-4E8B-A103-7D1845A21587}.Release|Mixed Platforms.Build.0 = Release|Any CPU
-		{85EE7B43-D2BF-4E8B-A103-7D1845A21587}.Release|Win32.ActiveCfg = Release|Any CPU
-		{85EE7B43-D2BF-4E8B-A103-7D1845A21587}.Release|x64.ActiveCfg = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-EndGlobal
