--- DesktopEditor/doctrenderer/test_builder/docbuilder/doctrendererwrapper/Stdafx.cpp
@@ -1,5 +1,0 @@
-// stdafx.cpp : source file that includes just the standard includes
-// doctrendererwrapper.pch will be the pre-compiled header
-// stdafx.obj will contain the pre-compiled type information
-
-#include "stdafx.h"
