--- DesktopEditor/doctrenderer/test_builder/docbuilder/doctrendererwrapper/Stdafx.h
@@ -1,7 +1,0 @@
-// stdafx.h : include file for standard system include files,
-// or project specific include files that are used frequently,
-// but are changed infrequently
-
-#pragma once
-
-
