--- DesktopEditor/doctrenderer/test_builder/docbuilder/doctrendererwrapper/resource.h
@@ -1,3 +1,0 @@
-//{{NO_DEPENDENCIES}}
-// Microsoft Visual C++ generated include file.
-// Used by app.rc
